requirements.txt
python-telegram-bot==13.15
APScheduler==3.6.3
pandas==2.2.2 